// web/src/service/types.ts
export type LanguageType = {
  id?: string
  name: string
  speaking_level: string
  writing_level: string
}